refactor code gen tool. closes #6
doesnt actually generate valid python code yet, but close enough for now
