+++ run_ltlsynt.sh
@@ -7,8 +7,8 @@ IN=$(syfco $filename.tlsf --print-input-signals)
 OUT=$(syfco $filename.tlsf --print-output-signals)
 ltlsynt --formula="$LTL" --ins="$IN" --outs="$OUT" > $filename.hoa
 sed -i 1d $filename.hoa
-./hoa2code $filename.hoa --arrow > $filename.code
-./parsehoa $filename.hoa > $filename.parsed.hoa
-autfilt $filename.parsed.hoa --deterministic --dot > $filename.dot
-sed -i 's/|/|\\n/g' $filename.dot
-dot -Tpng $filename.dot -o $filename.png
+./hoa2code $filename.hoa --python > $filename.code
+#./parsehoa $filename.hoa > $filename.parsed.hoa
+#autfilt $filename.parsed.hoa --deterministic --dot > $filename.dot
+#sed -i 's/|/|\\n/g' $filename.dot
+#dot -Tpng $filename.dot -o $filename.png
